requirements.txt
python-mnist==0.7
numpy==1.24.4